In the middle of implementing the comparisons.
